+++ xmake.lua
@@ -1112,6 +1112,10 @@ ${reset}]])
             os.setenvs(old_env)
             os.cd(path.join(prj_dir, "tests", "test_wave_vpi"))
             run_cmd("xmake run -P .")
+
+            os.setenvs(old_env)
+            os.cd(path.join(prj_dir, "tests", "test_wave_vpi_x"))
+            run_cmd("xmake run -P .")
         end
 
         --